Tests: remove obsolete JvmVersionRequirementTest
All test cases except `nestedClassMembers.kt` were already checked in
kotlinp tests.

Also, remove unused `LoadDescriptorUtil` and obsolete
`RecursiveDescriptorComparatorAdaptor`.
