--- compiler/testData/versionRequirement/patchVersion.kt
@@ -1,7 +1,0 @@
-@file:Suppress("INVISIBLE_MEMBER", "INVISIBLE_REFERENCE")
-package test
-
-import kotlin.internal.RequireKotlin
-
-@RequireKotlin("1.1.50", level = DeprecationLevel.HIDDEN)
-class Klass
