+++ compiler/tests/org/jetbrains/kotlin/serialization/klib/KotlinKlibSerializerTest.kt
@@ -21,15 +21,15 @@ import org.jetbrains.kotlin.cli.metadata.KotlinMetadataCompiler
 import org.jetbrains.kotlin.codegen.forTestCompile.ForTestCompileRuntime
 import org.jetbrains.kotlin.config.LanguageVersionSettingsImpl
 import org.jetbrains.kotlin.descriptors.impl.ModuleDescriptorImpl
-import org.jetbrains.kotlin.jvm.compiler.LoadDescriptorUtil.TEST_PACKAGE_FQNAME
 import org.jetbrains.kotlin.library.loader.KlibLoader
 import org.jetbrains.kotlin.library.metadata.KlibMetadataFactories
 import org.jetbrains.kotlin.library.metadata.NullFlexibleTypeDeserializer
+import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.storage.LockBasedStorageManager
 import org.jetbrains.kotlin.test.CompilerTestUtil
 import org.jetbrains.kotlin.test.TestCaseWithTmpdir
+import org.jetbrains.kotlin.test.services.JUnit5Assertions
 import org.jetbrains.kotlin.test.util.RecursiveDescriptorComparator
-import org.jetbrains.kotlin.test.util.RecursiveDescriptorComparatorAdaptor
 import org.junit.jupiter.api.Test
 import java.io.File
 
@@ -79,10 +79,11 @@ class KotlinKlibSerializerTest : TestCaseWithTmpdir() {
     private fun compareDumps(klibFile: File, source: String, goldenDataExtension: String) {
         val module = deserializeKlibToCommonModule(klibFile)
 
-        RecursiveDescriptorComparatorAdaptor.validateAndCompareDescriptorWithFile(
-            module.getPackage(TEST_PACKAGE_FQNAME),
+        RecursiveDescriptorComparator.validateAndCompareDescriptorWithFile(
+            module.getPackage(FqName("test")),
             RecursiveDescriptorComparator.DONT_INCLUDE_METHODS_OF_OBJECT,
-            File(source.replace(".kt", goldenDataExtension))
+            File(source.replace(".kt", goldenDataExtension)),
+            JUnit5Assertions,
         )
     }
